move layout views and constraints function to extension file in Views
